+++ adapter/ph/src/agent_output_worker.rs
@@ -1,4 +1,5 @@
 use crate::assembly::Assembly;
+use crate::config;
 use crate::counters_enum::*;
 use crate::fastpath;
 use crate::net_defs;
@@ -11,9 +12,6 @@ use zpr_ext::tokio_tun::*;
 pub struct Config {
     pub batch_size: usize,
 }
-
-// How much space to leave for the ZDP headers.
-const OUTBOUND_PACKET_HEADROOM: usize = 256;
 
 fn is_ip(pi: tun_pi::TunPi) -> bool {
     pi.proto == net_defs::ethertype::IP || pi.proto == net_defs::ethertype::IPV6
@@ -32,7 +30,7 @@ async fn worker(config: &Config, asm: &Assembly<'_>, tun: &Tun) {
         // since neither `read_buf()` nor `enqueue()` support it
         for mut buf in bufs.drain(..) {
             let pkt = loop {
-                let mut pkt = Packet::new(buf, OUTBOUND_PACKET_HEADROOM);
+                let mut pkt = Packet::new(buf, config::DEFAULT_MESSAGE_HEADROOM);
 
                 tun_recv_buf(tun, &mut pkt).await.unwrap();
                 let pi = tun_pi::read_pi(&mut pkt);
